qwen-image-edit/requirements.txt
torch>=2.0.0
git+https://github.com/huggingface/diffusers
transformers>=4.37.0
fastapi>=0.104.0
uvicorn>=0.24.0
pillow>=10.0.0
python-multipart>=0.0.6
pydantic>=2.4.0
requests>=2.31.0
numpy>=1.24.0
accelerate>=0.25.0
